Commit_20181025
Modify accept request method;
Complete product editer.

+++ demo/src/main/java/com/example/demo/mapper/ProductMapper.java
@@ -55,6 +55,6 @@ public interface ProductMapper {
     void update(Product product);
 
     @Delete("delete form product_table where id =#{id}")
-    void delete(Integer id);
+    void delete(long id);
 
 }
